separated serialization and deserialization into different files for ease of use, corrected documentation
